third_party/WebKit/LayoutTests/fast/css/hover-pseudo-element-quirks-expected.txt

Quirks mode - universal :hover on pseudo elements

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS window.eventSender is defined.
PASS getComputedStyle(t1, '::before').backgroundColor is transparent
PASS getComputedStyle(t1, '::before').backgroundColor is green
PASS getComputedStyle(t2, '::after').backgroundColor is transparent
PASS getComputedStyle(t2, '::after').backgroundColor is green
PASS successfullyParsed is true

TEST COMPLETE
